+++ src/app/(with-header)/layout.tsx
@@ -1,5 +1,4 @@
 import dynamic from "next/dynamic";
-import { cookies } from "next/headers";
 
 import { dehydrate, HydrationBoundary } from "@tanstack/react-query";
 
@@ -7,7 +6,7 @@ import { Footer } from "@/components/Footer/Footer";
 import { Header } from "@/components/Header/Header";
 import { memberKeys, memberQueries } from "@/features/member/hooks/useMemberHooks";
 import { MAIN_SCROLL_ID } from "@/hooks/useBodyScrollLock";
-import { ACCESS_TOKEN_COOKIE, REFRESH_TOKEN_COOKIE } from "@/lib/auth/cookie-constants";
+import { hasAuthCookies } from "@/lib/auth/hasAuthCookies";
 import { getQueryClient } from "@/lib/getQueryClient";
 
 const OnboardingModalWrapper = dynamic(() =>
@@ -29,13 +28,8 @@ export default async function Layout({ children }: { children: React.ReactNode }
   const queryClient = getQueryClient();
   let isAuthenticated = false;
   let memberProfile = null;
-  const cookieStore = await cookies();
-  const hasAuthCookie = Boolean(
-    cookieStore.get(ACCESS_TOKEN_COOKIE)?.value || cookieStore.get(REFRESH_TOKEN_COOKIE)?.value
-  );
-
   // 인증 쿠키가 없으면 서버에서 me 조회를 건너뛰어 초기 문서 TTFB를 줄인다.
-  if (hasAuthCookie) {
+  if (await hasAuthCookies()) {
     try {
       const response = await queryClient.fetchQuery(memberQueries.me());
       memberProfile = response.result;
