+++ src/app/(with-header)/session/[sessionId]/page.tsx
@@ -1,6 +1,5 @@
 import type { Metadata } from "next";
 
-import { cookies } from "next/headers";
 import { redirect } from "next/navigation";
 
 import { dehydrate, HydrationBoundary } from "@tanstack/react-query";
@@ -9,7 +8,7 @@ import { sessionApi } from "@/features/session/api";
 import { SessionPageContent } from "@/features/session/components/SessionPageContent";
 import { sessionQueries } from "@/features/session/hooks/useSessionHooks";
 import { isWaitingStatus } from "@/features/session/types";
-import { ACCESS_TOKEN_COOKIE, REFRESH_TOKEN_COOKIE } from "@/lib/auth/cookie-constants";
+import { hasAuthCookies } from "@/lib/auth/hasAuthCookies";
 import { getQueryClient } from "@/lib/getQueryClient";
 import { createPageMetadata } from "@/lib/seo/metadata";
 
@@ -49,12 +48,7 @@ export default async function SessionPage({ params }: SessionPageProps) {
     redirect(`/session/${sessionId}/waiting`);
   }
 
-  const cookieStore = await cookies();
-  const hasAuthCookie = Boolean(
-    cookieStore.get(ACCESS_TOKEN_COOKIE)?.value || cookieStore.get(REFRESH_TOKEN_COOKIE)?.value
-  );
-
-  if (hasAuthCookie) {
+  if (await hasAuthCookies()) {
     await queryClient.prefetchQuery(sessionQueries.waitingRoom(sessionId));
   }
