+++ src/features/lobby/components/LobbyHeader.tsx
@@ -1,17 +1,11 @@
 "use client";
 
-import { useCallback, useRef, useState } from "react";
+import { useRef } from "react";
 
 import { Button } from "@/components/Button/Button";
-import { AlertIcon } from "@/components/Icon/AlertIcon";
 import { ChevronLeftIcon } from "@/components/Icon/ChevronLeftIcon";
-import { Portal } from "@/components/Portal/Portal";
-import { useLeaveSession } from "@/features/session/hooks/useSessionHooks";
-import { useBodyScrollLock } from "@/hooks/useBodyScrollLock";
-import { ApiError } from "@/lib/api/api-client";
-import { DEFAULT_API_ERROR_MESSAGE } from "@/lib/error/error-codes";
-import { navigateWithHardReload } from "@/lib/navigation/hardNavigate";
-import { toast } from "@/lib/toast";
+import { LeaveConfirmDialog } from "@/features/session/components/LeaveConfirmDialog";
+import { useLeaveSessionHandler } from "@/features/session/hooks/useLeaveSessionHandler";
 
 interface LobbyHeaderProps {
   sessionId: string;
@@ -28,26 +22,12 @@ export function LobbyHeader({
   onCloseDialog,
   isLeavingRef,
 }: LobbyHeaderProps) {
-  const [serverError, setServerError] = useState<string | null>(null);
-  const leaveSessionMutation = useLeaveSession();
   const leaveButtonRef = useRef<HTMLButtonElement>(null);
-
-  const handleLeave = async () => {
-    setServerError(null);
-    isLeavingRef.current = true;
-
-    try {
-      await leaveSessionMutation.mutateAsync({ sessionRoomId: sessionId });
-      onCloseDialog();
-      // 하드 네비게이션으로 캐시 클리어 및 SSE 연결 정리
-      navigateWithHardReload("/");
-    } catch (error: unknown) {
-      isLeavingRef.current = false;
-      const message = error instanceof ApiError ? error.message : DEFAULT_API_ERROR_MESSAGE;
-      setServerError(message);
-      toast.error(message);
-    }
-  };
+  const { handleLeave, isPending, serverError } = useLeaveSessionHandler({
+    sessionId,
+    isLeavingRef,
+    onCloseDialog,
+  });
 
   return (
     <>
@@ -83,91 +63,16 @@ export function LobbyHeader({
 
       {showDialog && (
         <LeaveConfirmDialog
+          description="나가기를 하면 대기 방에서 나가게 됩니다."
           onClose={() => {
             onCloseDialog();
             leaveButtonRef.current?.focus();
           }}
           onConfirm={handleLeave}
-          isPending={leaveSessionMutation.isPending}
+          isPending={isPending}
           serverError={serverError}
         />
       )}
     </>
-  );
-}
-
-interface LeaveConfirmDialogProps {
-  onClose: () => void;
-  onConfirm: () => void;
-  isPending: boolean;
-  serverError: string | null;
-}
-
-function LeaveConfirmDialog({
-  onClose,
-  onConfirm,
-  isPending,
-  serverError,
-}: LeaveConfirmDialogProps) {
-  useBodyScrollLock();
-
-  const dialogRef = useRef<HTMLDialogElement | null>(null);
-
-  const setDialogRef = useCallback((node: HTMLDialogElement | null) => {
-    if (node && !node.open) {
-      node.showModal();
-    }
-    dialogRef.current = node;
-  }, []);
-
-  const handleBackdropClick = (event: React.MouseEvent<HTMLDialogElement>) => {
-    if (event.target !== dialogRef.current) return;
-    onClose();
-  };
-
-  return (
-    <Portal>
-      <dialog
-        ref={setDialogRef}
-        onCancel={onClose}
-        onClick={handleBackdropClick}
-        className="bg-surface-default p-3xl fixed inset-0 m-auto flex w-full max-w-80 flex-col gap-6 rounded-lg border border-gray-900 backdrop:bg-(--color-overlay-default)"
-      >
-        <div className="flex flex-col gap-2">
-          <h2 className="text-text-primary text-lg font-bold">세션에서 나가시겠어요?</h2>
-          <p className="text-text-secondary text-sm">나가기를 하면 대기 방에서 나가게 됩니다.</p>
-        </div>
-
-        {serverError && (
-          <div className="flex animate-[fadeIn_0.2s_ease-out] items-center gap-2 rounded-sm bg-red-500/10 px-4 py-3 text-sm text-red-500">
-            <AlertIcon className="h-4 w-4 shrink-0" />
-            {serverError}
-          </div>
-        )}
-
-        <div className="flex w-full gap-2">
-          <Button
-            variant="ghost"
-            colorScheme="secondary"
-            size="medium"
-            className="flex-1"
-            onClick={onClose}
-            disabled={isPending}
-          >
-            취소
-          </Button>
-          <Button
-            variant="solid"
-            colorScheme="primary"
-            size="medium"
-            className="flex-1"
-            onClick={onConfirm}
-            disabled={isPending}
-          >
-            {isPending ? "나가는 중..." : "나가기"}
-          </Button>
-        </div>
-      </dialog>
-    </Portal>
   );
 }
